% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exposure2OutcomeSurv-package.R
\docType{package}
\name{Exposure2OutcomeSurv-package}
\alias{Exposure2OutcomeSurv}
\alias{Exposure2OutcomeSurv-package}
\title{Exposure2OutcomeSurv: Survival Analysis on Exposure and Outcome Condition Pairs}
\description{
Shiny application for exploring exposure-outcome survival relationships using matched cohort data from an OMOP CDM. Uses Kaplan-Meier plots for visualisation and the Cox model to calculate Hazard Ratios.
}
\author{
\strong{Maintainer}: Sander Soodla \email{sandersoodla@gmail.com}

}
\keyword{internal}
